Created Java Project in Eclipse
Currently researching how to use EGit and Git in general, not sure if I
did this right lol. Anyway, the project has been created, though it was
originally missing a src folder and I had to add that manually, even
though I'm pretty sure that there's a way to do that automatically. The
first steps have been taken!